// src/Model/iTechRestApi/iTechControlTableReferenceEnum.ts
//*************************DO NOT MODIFY**************************
//
//THESE FILES ARE AUTOGENERATED WITH A TYPESCRIPT CODE GENERATION TOOL AND ANY MODIFICATIONS MADE HERE WILL BE LOST.
//PLEASE VISIT http://frhagn.github.io/Typewriter/ TO LEARN MORE ABOUT THIS VISUAL STUDIO EXTENSION
//
//*************************DO NOT MODIFY**************************



export enum iTechControlTableReferenceEnum {
    iTechDataCase = 1,
    iTechDataTask = 2,
    iTechDataNote = 3,
    iTechDataSecurityObject = 4,
    iTechAudit = 5,
    iTechHclWebSphere = 6,
    iTechSim = 7,
    iTechSimAccident = 8,
    iTechDataResultSet = 9,
    iTechSimSalesForce = 10,
    iTechDataCaseToDatabase = 11,
    termToReference = 12,
    iTechDataTerm = 13,
    iTechSimUpload = 14,
    areaToReference = 15,
    unredactToReference = 16,
    iTechDataRule = 17,
    iTechDataAlert = 18,
    iTechStockHeader = 19,
    iTechStockOrderManagement = 20,
    iTechStockAimAudt = 21,
    iTechDataSurveillance = 22,
    iTechDataTag = 23,
    iTechDataLexicon = 24,
    iTechDataWebReportConfiguration = 25,
    iTechDataWebTab = 26,
    iTechDataLexiconGroup = 27,
    iTechStockOrder = 28
}